# api/grpc/Perun_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "grpc/Perun_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PerunAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/service/run": {
      "post": {
        "operationId": "PerunAPI_RunService",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/perun_apiRunServiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/perun_apiRunServiceRequest"
            }
          }
        ],
        "tags": [
          "PerunAPI"
        ]
      }
    },
    "/velez": {
      "post": {
        "operationId": "PerunAPI_ConnectVelez",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/perun_apiConnectVelezResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/perun_apiConnectVelezRequest"
            }
          }
        ],
        "tags": [
          "PerunAPI"
        ]
      }
    },
    "/velez/list": {
      "post": {
        "operationId": "PerunAPI_ListNodes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/perun_apiListNodesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/perun_apiListNodesRequest"
            }
          }
        ],
        "tags": [
          "PerunAPI"
        ]
      }
    },
    "/version": {
      "get": {
        "operationId": "PerunAPI_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/perun_apiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "PerunAPI"
        ]
      }
    }
  },
  "definitions": {
    "perun_apiConnectVelezRequest": {
      "type": "object",
      "properties": {
        "node": {
          "$ref": "#/definitions/perun_apiNode"
        },
        "ssh": {
          "$ref": "#/definitions/perun_apiSsh"
        }
      }
    },
    "perun_apiConnectVelezResponse": {
      "type": "object"
    },
    "perun_apiListNodesRequest": {
      "type": "object",
      "properties": {
        "searchPattern": {
          "type": "string"
        },
        "paging": {
          "$ref": "#/definitions/perun_apiListPaging"
        }
      }
    },
    "perun_apiListNodesResponse": {
      "type": "object",
      "properties": {
        "nodes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/perun_apiNode"
          }
        }
      }
    },
    "perun_apiListPaging": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "format": "int64"
        },
        "offset": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "perun_apiNode": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "addr": {
          "type": "string"
        },
        "customVelezKeyPath": {
          "type": "string"
        },
        "securityDisabled": {
          "type": "boolean"
        }
      }
    },
    "perun_apiRunServiceRequest": {
      "type": "object",
      "properties": {
        "imageName": {
          "type": "string"
        },
        "serviceName": {
          "type": "string"
        },
        "replicationFactor": {
          "type": "integer",
          "format": "int64",
          "title": "the amount of replicas needed to start and support running"
        }
      }
    },
    "perun_apiRunServiceResponse": {
      "type": "object"
    },
    "perun_apiSsh": {
      "type": "object",
      "properties": {
        "keyBase64": {
          "type": "string",
          "format": "byte"
        },
        "addr": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "perun_apiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}